bash: change break alias from break to br
I was not able to figure out cause, but when I type some command and
press tab to using auto complete feature in bash, break command
accidently triggerd. This is caused when `yay -R<tab>` and some of other
commands.

I just simply changed the alias to br, and the bug has been disappeared